feat(frontend): Add User Authentication
This includes more than I'd like, but the effort to remove some of
the code wasn't worth the effort.

It does add a Navbar to the top of the page, with buttons to manage
the register/login/logout features using `next-auth`. I have added
2 social logins `GitHub` and `Google`, as well as the ability to
register with a username/password (username is actually the email).