feat(frontend): Add ability to filter listings
To make it easier for a user to find listings of interest, I've
added the ability to filter listings by a number of features e.g.
location, no of rooms/guests/bathrooms etc. The filters are added
to the URL, so the user can copy/paste that to another user so they
too could narrow the list to what they needed.